fix clipboard with when
